requirements.txt
google-genai
pyaudio
mss
pyscreenshot
pillow
python-dotenv
